Change optimal exchange and display for each players
The best exchange is the one that exchange the most value for each player,
right now as well as potentially.

We don't want to take time to exchange cheap objects that we may not keep
later anyway. Once a stack with very pricey object is full we don't have to
bother with it again.