Optimize the computation for BAG_UPDATE
This is problematic when we add something to the scrap junk
list. There is not event to listen to this, and as there is
a lot less calculation of the cheapest object so the problem
starts to be really apparent. (Added calculation when
triggerring manually so it's fixable easily)

Fix #28